Nearest neighbors from a file.

Compute nearest distance to each phenotype for each cell in a (possibly merged) inForm cell seg table. Add Distance to <phenotype> columns. Write the result to a new file.

compute_all_nearest_distance(cell_table_path = NULL, out_path = NULL)

Arguments

cell_table_path

Path to an inForm cell seg data file, or NULL to prompt for the path.

out_path

Path to the output file, or NULL to create a path from the input file path.

Details

NOTE: The input file is read using read_cell_seg_data so the conversions and cleanup it does will be applied to the output data.

See also