Plot a distance map using a diverging color scale with white at 0

Plot a distance map using a diverging color scale with white at 0

plot_diverging(im, title, show_boundary = FALSE, ...)

Arguments

im
A pixel image of class im.

title
Title for the plot.

show_boundary
Should the boundary be highlighted?

...
Additional arguments passed to plot.im