Count cells within a radius

Count the number of from cells having a to cell within radius microns in tissue category category. Compute the average number of to cells within radius of from cells.

count_within(csd, from, to, radius, category = NULL, dst = NULL)

Arguments

csd
A data frame with Cell X Position, Cell Y Position and Phenotype columns, such as the result of calling read_cell_seg_data.

from, to
Selection criteria for the rows and columns. Accepts all formats accepted by select_rows.

radius
The radius or radii to search within.

category
Optional tissue category to restrict both from and to.

dst
Optional distance matrix corresponding to csd, produced by calling distance_matrix.

Value

A data_frame with four columns and one row for each value in radius:

radius

The value of radius for this row.

from_count

The number of from cells found in csd.

to_count

The number of to cells found in csd.

from_with

The number of from cells with a to cell within radius.

within_mean

The average number of to cells found within radius microns of each from cell.

Details

For each from cell, count the number of to cells within radius microns. Report the number of from cells containing at least one to cell within radius as from_with. Report the average number of to cells as within_mean.

If category is specified, all reported values are for cells within the given tissue category. If category is NULL, values are reported for the entire data set. radius may be a vector with multiple values.

See also

Other distance functions: compute_all_nearest_distance, distance_matrix, find_nearest_distance, subset_distance_matrix