Count cells within a radius.

Count the number of from cells having a to cell within radius microns in tissue category category. Compute the average number of to cells within radius of from cells.

count_within(csd, from, to, radius, category = NA, dst = NULL)

Arguments

csd
A data frame with Cell X Position, Cell Y Position and Phenotype columns, such as the result of calling read_cell_seg_data.

from, to
Selection criteria for the rows and columns. Accepts all formats accepted by select_rows.

radius
The radius or radii to search within.

category
Optional tissue category to restrict both from and to.

dst
Optional distance matrix corresponding to csd, produced by calling distance_matrix.

Value

A data_frame with five columns and one row for each value in radius:

radius

The value of radius for this row.

from_count

The number of from cells found in csd.

to_count

The number of to cells found in csd.

from_with

The number of from cells with a to cell within radius.

within_mean

The average number of to cells found within radius microns of each from cell.

Details

For each from cell, count the number of to cells within radius microns. Report the number of from cells containing at least one to cell within radius as from_with. Report the average number of to cells per from cell as within_mean.

There are some subtleties to this calculation. See the example code for a demonstration.

It is not symmetric in from and to. For example the number of tumor cells with a macrophage within 25 microns is not the same as the number of macrophages with a tumor cell within 25 microns.

from_count*within_mean is not the number of to cells within radius of a from cell, it may count to cells multiple times.

Surprisingly, from_count*within_mean is symmetric in from and to. The double-counting works out.

To aggregate within_mean across multiple samples (e.g. by Slide ID) use code such as

results %>% group_by(`Slide ID`, from, to, radius) %>%
  summarize(from_count=sum(from_count),
            within=sum(from_count*within_mean),
            avg=within/from_count) %>%
  ungroup 

If category is specified, all reported values are for cells within the given tissue category. If category is NA, values are reported for the entire data set. radius may be a vector with multiple values.

See also

Other distance functions: compute_all_nearest_distance, count_touching_cells, count_within_batch, distance_matrix, find_nearest_distance, spatial_distribution_report, subset_distance_matrix

Examples

library(tidyverse)
csd = sample_cell_seg_data

# Find the number of macrophages with a tumor cell within 10 or 25 microns
count_within(csd, from='CD68+', to='CK+', radius=c(10, 25)) %>%
  mutate(not_to=from_count*within_mean, to_mean=not_to/to_count)

#> # A tibble: 2 x 7
#>   radius from_count to_count from_with within_mean not_to    to_mean
#>    <dbl>      <int>    <int>     <int>       <dbl>  <dbl>      <dbl>
#> 1     10        417     2257        89   0.2877698    120 0.05316792
#> 2     25        417     2257       274   3.2781775   1367 0.60567125

# Find the number of tumor cells with a macrophage within 10 or 25 microns
# Show that from_count*within_mean is not the same as from_with in the
# previous computation but from_count*within_mean/to_count is the same.
count_within(csd, from='CK+', to='CD68+', radius=c(10, 25)) %>%
  mutate(not_to=from_count*within_mean, to_mean=not_to/to_count)

#> # A tibble: 2 x 7
#>   radius from_count to_count from_with within_mean not_to   to_mean
#>    <dbl>      <int>    <int>     <int>       <dbl>  <dbl>     <dbl>
#> 1     10       2257      417       109  0.05316792    120 0.2877698
#> 2     25       2257      417       664  0.60567125   1367 3.2781775