Nearest neighbor distances for each cell and phenotype.

For each phenotype in a single sample, find the distance from each cell to the nearest other cell in the phenotype.

find_nearest_distance(csd, phenotypes = NULL)

Arguments

csd
A data frame with Cell X Position, Cell Y Position and Phenotype columns, such as the result of calling read_cell_seg_data.

phenotypes
Optional list of phenotypes to include. If omitted, unique(csd$Phenotype) will be used.

Value

A data_frame containing a Distance to <phenotype> column for each phenotype. Will contain NA values where there is no other cell of the phenotype.

See also

compute_all_nearest_distance which applies this function to a (possibly merged) data file.

Other distance functions: compute_all_nearest_distance, count_touching_cells, count_within_batch, count_within, distance_matrix, spatial_distribution_report, subset_distance_matrix

Examples

# Compute distance columns
csd = sample_cell_seg_data
nearest = find_nearest_distance(csd)
dplyr::glimpse(nearest)

#> Observations: 5,726
#> Variables: 6
#> $ Distance to cytotoxic CD8   <dbl> 15.231546, 21.505813, 14.008926, 11.800...
#> $ Distance to helper CD4      <dbl> 140.12584, 136.88042, 137.29530, 138.88...
#> $ Distance to macrophage CD68 <dbl> 11.597414, 28.160256, 9.552487, 24.6829...
#> $ Distance to other           <dbl> 67.742158, 56.147128, 65.431262, 80.188...
#> $ Distance to T reg Foxp3     <dbl> 504.7091, 495.7673, 501.5187, 510.5331,...
#> $ Distance to tumor           <dbl> 3.201562, 9.552487, 3.201562, 13.756816...

# Make a combined data frame including original data and distance columns
csd = cbind(csd, find_nearest_distance(csd))
# NOT RUN {
# If `merged` is a data frame containing cell seg data from multiple fields,
# this code will create a new `data_frame` with distance columns computed
# for each `Sample Name` in the data.
merged_with_distance = merged %>%
  dplyr::group_by(`Sample Name`) %>%
  dplyr::do(dplyr::bind_cols(., find_nearest_distance(.)))
# }